+++ test/solidlsp/latex/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for LaTeX language server functionality."""